refactor: move resync up a level out of v1
