+++ cognite/powerops/resync/__init__.py
@@ -1,7 +1,3 @@
-from .v1 import apply as apply_v1
-from .v1 import plan as plan_v1
+from .main import apply, plan
 
-__all__ = [
-    "plan_v1",
-    "apply_v1",
-]
+__all__ = ["plan", "apply"]
