--- cognite/powerops/resync/v1/__init__.py
@@ -1,3 +1,0 @@
-from .main import apply, plan
-
-__all__ = ["plan", "apply"]
